Imports activity events data

Imports the CSV (semi-colon separted value) activPAL EventsEX file for the subject and trimester of interest that was exported with the activPAL software.

Usage

read_events(subject, trimester, file = NULL, ...)

Arguments

subject

A character vector denoting the Pregnancy 24/7 subject.

trimester

A character or integer vector denoting trimester of pregnancy.

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). The default value is NULL, which uses the subject and trimester parameters to find the appropriate filename.

...

Any other parameters needed for utils::read.csv().

Value

A data.events object is returned, which is a data.frame that only includes the raw activPAL EventsEx data for the subject of interest during the trimester of interest.

Details

When file is the default value of NULL the subject and trimester parameters along with the current working directory are used to find the file that ends in EventsEx.csv by using the loc() function. As the purpose of this function is to work in the background of the process_data() function, a filename should only be specified when read_events() is used outside of the process_data(), usually for individual examination of problematic data.

By default, the column names are skipped as the activPAL software tends to produce variable names that do NOT conform to the standard R naming convention.

See also

Examples

if (FALSE) { # \dontrun{
# numeric trimester
dat <- read_events("0001-AB", 3)
} # }

if (FALSE) { # \dontrun{
# charactter trimester
dat <- read_events("0001-AB", "3")
} # }