Summarize daily activity

The activities during the wear period are summarized using events and 1 second EPOCH data for the subject of interest during the trimester of interest.

Usage

summarize_daily(
  subject,
  trimester,
  data_events,
  data_1s_epoch,
  all_days = 1:9,
  good_days = 1:9
)

Arguments

subject

A character vector denoting the Pregnancy 24/7 subject.

trimester

A character or integer vector denoting trimester of pregnancy.

data_events

Either a process.data.events or merge.data.events data set.

data_1s_epoch

A one.epoch.data object created by create_1s_epoch().

all_days

An integer vector denoting the all the possible wear days. The default is 1:9, as for the Pregnancy 24/7 study the wear period last nine days.

good_days

An integer vector denoting the wear days that have valid data. The default is 1:9, as for the Pregnancy 24/7 study the wear period last nine days.

Value

A named list with objects:

daily a data.frame containing daily summary of activities for all the possible wear days

avgs a data.frame containing the averages of the activities during the wear period.

Details

See the vingettes for a full description of the variables created.

Examples

if (FALSE) { # \dontrun{
process_dat <- process_events(data = dat)
sec_by_sec <- create_1s_epoch(data = process_dat)
summ_daily <- summarize_daily(
 subject = "0001-AB", trimester = "1",
 data_events = process_dat, data_1s_epoch = sec_by_sec
)
} # }